added docker support, dockerfile, changed app routes, and /app structure. Also cleaned up whole project, namely file structure and organiation
