+++ app/requirements.txt
@@ -1,0 +1,8 @@
+flask
+flask-cors
+joblib
+pandas
+xgboost
+numpy
+python-dotenv
+requests
