+++ app/templates/index.html
@@ -1,7 +1,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Forecasting NYISO</title>
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="../static/style.css">
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.9.1/chart.min.js"></script>
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
@@ -52,5 +52,5 @@
     </div>
   </div>
 
-  <script src="frontend.js"></script>
+  <script src="../static/frontend.js"></script>
 </body>
