+++ deprecated/ARIMA.py
