+++ deprecated/XGDaily.py
