+++ deprecated/tools/autoARIMA.py
