+++ deprecated/useARIMA.py
