--- guide.md
@@ -1,44 +1,0 @@
-1. Exploratory Data Analysis (EDA)
-First, let's understand the data you already have:
-
-Examine the time series structure (missing values, outliers)
-Analyze seasonality patterns (daily, weekly, annual)
-Look for trends and recurring patterns
-Visualize autocorrelation to understand temporal dependencies
-
-2. Feature Engineering
-Rather than just collecting more data initially, focus on extracting useful features:
-
-Calendar features (hour, day of week, month, holidays)
-Lagged variables (previous day, previous week)
-Rolling statistics (moving averages, standard deviations)
-Cyclical encodings of time variables
-
-3. External Data Collection
-Now consider what additional data might improve predictions:
-
-Weather data (temperature, humidity, etc.)
-Economic indicators
-Special events data
-COVID-19 impact periods
-Demographic changes
-
-4. Model Development Strategy
-I'd recommend a staged approach:
-
-Baseline models (ARIMA, ETS, simple regression)
-Advanced models (Prophet, LSTM, XGBoost)
-Ensemble methods combining multiple approaches
-
-5. Evaluation Framework
-
-Use appropriate metrics (RMSE, MAPE, etc.)
-Establish proper validation strategy (walk-forward validation)
-Compare to simple baseline models
-Test on multiple forecast horizons (next day, next week, next month)
-
-6. Deployment Planning
-
-Determine update frequency
-Establish retraining schedule
-Design monitoring system for model drift
